sanity check plots
Integral of simple equation matches the manually adjusted split integral
of the Gdot equation... however, it is not really clear what the optimal
singularity delta may be. Currently it is 1E-5, but it may be more
prudent for this to be adaptive... i.e., reduce delta by order of
magnitude until integration warning goes away? Maybe this is too crude
though...