Split integral should be sufficient
You can see that the spread of the discontinuity is approximately the
same for low values of `t`. This means that you can rely on the minus
1e-5 split integration for small values of `t`.

TODO: However, for large
values of t, you see that to the right of the discontinuity you get
positive values, so you should verify this by messing around a little
bit with singularity delta for large `t`.